change query to update margin
